Added put at /post/:id/likePost

+++ server/controllers/posts.js
@@ -1,4 +1,5 @@
 const Post = require('../models/post')
+const mongoose = require('mongoose');
 
 const getPosts = async(req,res)=>{
     try {
@@ -23,8 +24,24 @@ const createPost = async(req,res) => {
 const updatePost = async(req,res) => {
     try {
         const {id} = req.params;
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(404).json({"message": "Post Not Found."});
+        }
         const {title,message,creator,tags,selectedFile} = req.body;
         const updatedPost = await Post.findByIdAndUpdate(id,{title,message,creator,tags,selectedFile},{new:true})
+        res.status(201).json({data:updatedPost})
+    } catch (error) {
+        res.status(500).json({message:error})
+    }
+}
+const likePost = async(req,res) => {
+    try {
+        const {id} = req.params;
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(404).json({"message": "Post Not Found."});
+        }
+        const post = await Post.findById(id);
+        const updatedPost = await Post.findByIdAndUpdate(id,{likeCount:post.likeCount+1},{new:true})
         res.status(201).json({data:updatedPost})
     } catch (error) {
         res.status(500).json({message:error})
@@ -34,6 +51,9 @@ const updatePost = async(req,res) => {
 const deletePost = async(req,res) => {
     try {
         const {id} = req.params;
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(404).json({"message": "Post Not Found."});
+        }
         await Post.findByIdAndDelete(id)
         res.status(201).json({data:"Deleted Successfully."})
     } catch (error) {
@@ -42,4 +62,6 @@ const deletePost = async(req,res) => {
 }
 
 
-module.exports = {getPosts,createPost,updatePost,deletePost}
+
+
+module.exports = {getPosts,createPost,updatePost,deletePost,likePost}
